% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_seats.R
\name{allocate_seats_nl}
\alias{allocate_seats_nl}
\title{Allocate seats per the Dutch Parliamentary electoral system.}
\usage{
allocate_seats_nl(parties, votes)
}
\arguments{
\item{parties}{vector of party names}

\item{votes}{vector of vote counts, of the same length as \code{parties}}
}
\value{
see \code{allocate_votes}
}
\description{
This is a thin wrapper around \code{allocate_seats(...)} using default settings as
used in Dutch for the Dutch house of representatives ("Tweede Kamer").
}
