% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disproportionality.R
\name{gallagher_index}
\alias{gallagher_index}
\title{Gallagher index of disproportionality}
\usage{
gallagher_index(votes, seats)
}
\arguments{
\item{votes}{vector of votes for each party}

\item{seats}{vector of seats allocated to each party}
}
\value{
a value between 0 and 100
}
\description{
Given two vectors of vote counts and allocated seats, calculate the Gallagher
(or Least Squares) index of disproportionality. This index takes a middle ground
between the Saint-Langue and Loosemore Hanby indices, that are minimized by
largest averages and largest remainder seat allocation methods, respectively.
}
\details{
A Gallagher index of zero corresponds to perfectly proportional seat allocation,
whereas a high index indicates greater disproportionality. The Gallagher index
ranges from 0 to 100.
}
