% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disproportionality.R
\name{loosemore_hanby_index}
\alias{loosemore_hanby_index}
\title{Loosemore-Hanby index of disproportionality}
\usage{
loosemore_hanby_index(votes, seats)
}
\arguments{
\item{votes}{vector of votes for each party}

\item{seats}{vector of seats allocated to each party}
}
\value{
a value between 0 and 1.
}
\description{
Given two vectors of vote counts and allocated seats, calculate the Loosemore-Hanby
index of disproportionality. This index is minimized by largest remainder seat
allocation methods, e.g. Hare.
}
\details{
A Loosemore-Hanby index of zero corresponds to perfectly proportional seat allocation,
whereas a high index indicates greater disproportionality. The Loosemore-Hanby
index ranges from 0 to 1.
}
