% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional-package.R
\docType{package}
\name{regional-package}
\alias{regional}
\alias{regional-package}
\title{regional: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

}
\keyword{internal}
