% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disproportionality.R
\name{saint_lague_index}
\alias{saint_lague_index}
\title{Saint-Langue index of disproportionality}
\usage{
saint_lague_index(votes, seats)
}
\arguments{
\item{votes}{vector of votes for each party}

\item{seats}{vector of seats allocated to each party}
}
\value{
a value between 0 and infinity.
}
\description{
Given two vectors of vote counts and allocated seats, calculate the Saint-Langue
index of disproportionality. This index is minimized by largest averages seat
allocation methods, e.g. d'Hondt or (unsurprisingly) Saint-Langue.
}
\details{
A Saint-Langue index of zero corresponds to perfectly proportional seat allocation,
whereas a high index indicates greater disproportionality. The Saint-Langue index
is theoretically infinite if any party is allocated a seat without receiving
any votes.
}
