% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disproportionality.R
\name{saint_langue_index}
\alias{saint_langue_index}
\title{Saint-Langue index of disproportionality}
\usage{
saint_langue_index(votes, seats)
}
\arguments{
\item{votes}{vector of votes for each party}

\item{seats}{vector of seats allocated to each party}
}
\value{
a value between 0 and infinity.
}
\description{
Given two vectors of vote counts and allocated seats, calculate the Saint-Langue
index of disproportionality. This index is minimized by largest averages seat
allocation methods, e.g. d'Hondt or (unsurprisingly) Saint-Langue.
}
\details{
A Saint-Langue index of zero corresponds to perfectly proportional seat allocation,
whereas a high index indicates greater disproportionality. The Saint-Langue index
is theoretically infinite if any party is allocated a seat without receiving
any votes.
}
\examples{
# Saint-Langue index for a mostly proportional outcome
votes <- c(0.4, 0.3, 0.2, 0.06, 0.03, 0.01) * 1e6
parties <- paste("Party", letters[1:6])
outcome <- allocate_seats(parties, votes, 120)
saint_langue_index(votes, outcome$seats)

# Using Hare allocation
outcome_hare <- allocate_seats(parties, votes, 120, method = "hare")
saint_langue_index(votes, outcome_hare$seats)

# Saint-Langue is (over)sensitive to small parties gaining fixed seats
# assigning a fixed seat to a tiny party (e.g. to ensure minority representation)
votes <- c(0.409, 0.3, 0.2, 0.06, 0.03, 0.001) * 1e6
outcome <- allocate_seats(parties, votes, 120, fixed_seats = c(rep(0,5), 1))
saint_langue_index(votes, outcome$seats)

}
