% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_districts.R
\name{simulate_districts}
\alias{simulate_districts}
\title{Helper function to simulate district population size and turnout rates.}
\usage{
simulate_districts(
  districts,
  func_population = function(n_districts) rep(1.5e+07/n_districts, n_districts),
  func_turnout = function(n_districts) stats::runif(n_districts, 0.5, 0.7)
)
}
\arguments{
\item{districts, }{vector of district names or integer number of districts.}

\item{func_population, }{function called to get district population sizes, defaults
to an anonymous function that returns \code{15e6} (i.e., 15 million) equally distributed
over the number of districts. This number is chosen to be roughly analogous to the
total number of voters in the Netherlands. This function is called with a single
argument \code{n_districts}, and should return a vector of length \code{n_districts}.}

\item{func_turnout, }{function called to get district turnout rates. Defaults to
a uniform distribution with min = 0.5 and max = 0.7. This function is called with a single
argument \code{n_districts}, and should return a vector of length \code{n_districts}. Turnout rates
should be in the range 0-1.}
}
\value{
a tibble with three columns:
\item{district}{district name (or index)}
\item{district_population}{district population}
\item{district_turnout}{district turnout rate}
}
\description{
Helper function to simulate district population size and turnout rates.
}
