% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_vote_shares.R
\name{simulate_shares_dirichlet}
\alias{simulate_shares_dirichlet}
\title{Simulate vote shares based on a dirichlet distribution, using parameters
estimated by CITATION NEEDED}
\usage{
simulate_shares_dirichlet(n_parties)
}
\arguments{
\item{n_parties}{integer, number of parties in each district.}
}
\value{
a vector of vote shares of length \code{n_parties} that sums to 1.
}
\description{
Simulate vote shares based on a dirichlet distribution, using parameters
estimated by CITATION NEEDED
}
