% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_seats.R
\name{allocate_seats}
\alias{allocate_seats}
\title{Allocate seats across parties using different methods of allocation}
\usage{
allocate_seats(
  parties,
  votes,
  seats,
  fixed_seats = rep(0, length(parties)),
  method = c("dh", "hare"),
  threshold = 0,
  fixed_seat_overrides_threshold = FALSE
)
}
\arguments{
\item{parties}{vector of party names (or other identifiers)}

\item{votes}{vector of vote counts per party}

\item{seats}{total number of seats available, \strong{including} any seats already
allocated using the \code{fixed_seats} argument.}

\item{fixed_seats}{vector of seat counts already allocated to each party that
should be considered 'fixed'. Defaults to a zero vector, i.e., no seats have
yet been allocated.}

\item{method}{one of several methods of seat allocation. See details. Defaults
to 'dh' for d'Hondt method.}

\item{threshold}{vote share required to be eligible for any seats. Defaults to
0, should be in the range \link{0,1}. Common values are between 0 and 0.05.}

\item{fixed_seats_override_threshold}{should parties with fixed seats automatically
be eligible for remainder seat allocation? This is common in some parliamentary
systems. Defaults to \code{FALSE}.}
}
\value{
A tibble containing several columns:
\item{party}{party name}
\item{votes}{number of votes}
\item{seats}{total number of assigned seats}
\item{fixed_seats}{number of previously allocated seats considered 'fixed'}
\item{allocated_seats}{number of seats allocated in this allocation step,
equal to \code{seats - fixed_seats}}
\item{ideal_seats}{'ideal' number of seats under perfectly proportional
representation. Calculated as \code{votes / sum(votes) * seats}}
}
\description{
Given a set of parties and associated vote counts, allocate \code{seats} number of
seats using one of several common methods of seat allocation.
}
