% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_distinction.R
\name{reg_distinction}
\alias{reg_distinction}
\title{Distinction}
\usage{
reg_distinction(
  region,
  raster,
  dist_fun = "euclidean",
  sample_size = 1,
  unit = "log2",
  na.rm = FALSE,
  engine = c("lowmem", "speed"),
  ...
)
}
\arguments{
\item{region}{An object of class \code{sf} with a \code{POLYGON} or \code{MULTIPOLYGON} geometry type}

\item{raster}{An object of class SpatRaster (terra)}

\item{dist_fun}{Distance measure used. This function uses \code{philentropy::distance} (run \code{philentropy::getDistMethods()} to find possible distance measures) or \code{proxy::dist} (run \code{names(proxy::pr_DB$get_entries())} to find possible distance measures) in the background.
It is also possible to use \code{"dtw"} (dynamic time warping)}

\item{sample_size}{Proportion of the cells inside of each region to be used in calculations. Value between 0 and 1.
It is also possible to specify an integer larger than 1, in which case the specified number of cells
of each region will be used in calculations.}

\item{unit}{a character string specifying the logarithm unit that should be used to
compute distances that depend on log computations.}

\item{na.rm}{Whether NA values should be stripped from the calculations.}

\item{engine}{A character string specifying the computation strategy:
\code{"lowmem"} extracts values on the fly (lower memory, slower),
\code{"speed"} caches per-region extracts (higher memory, faster).}

\item{...}{Additional arguments for \code{philentropy::dist_one_one}, \code{proxy::dist}, or \code{dtwclust::dtw_basic}.
When \code{dist_fun = "dtw"} is used, \code{ndim} should be set to specify how many dimension the input raster time-series has.}
}
\value{
A vector with the distinction values
}
\description{
Distinction is an average distance between the focus region and all of the other regions.
This value is between 0 and 1, where large value indicates that the values in the region
stands out from the other regions.
}
\examples{
\dontrun{
  library(terra)
  if (requireNamespace("sf", quietly = TRUE)) {
    library(sf)
    volcano = rast(system.file("raster/volcano.tif", package = "regional"))
    vr = read_sf(system.file("regions/volcano_regions.gpkg", package = "regional"))
    vr$dis = reg_distinction(vr, volcano, sample_size = 0.5)

    mean(vr$dis)

    plot(volcano)
    plot(vect(vr), add = TRUE)
    plot(volcano)
    plot(vr["dis"], add = TRUE)
 }
}
}
