% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_distinction.R
\name{reg_distinction}
\alias{reg_distinction}
\title{Distinction}
\usage{
reg_distinction(region, raster, dist_fun = "euclidean", sample_size = 1)
}
\arguments{
\item{region}{An object of class \code{sf} with a \code{POLYGON} or \code{MULTIPOLYGON} geometry type}

\item{raster}{An object of class SpatRaster (terra)}

\item{dist_fun}{Distance measure used. This function uses the \code{philentropy::distance} function in the background. Run \code{philentropy::getDistMethods()} to find possible distance measures.}

\item{sample_size}{Proportion of the cells inside of each region to be used in calculations. Value between 0 and 1.}
}
\value{
A vector with the distinction values
}
\description{
Distinction is an average distance between the focus region and all of the other regions.
This value is between 0 and 1, where large value indicates that the values in the region
stands out from the other regions.
}
\examples{
\dontrun{
  library(terra)
  library(sf)
  volcano = rast(system.file("raster/volcano.tif", package = "supercells"))
  vr = read_sf(system.file("regions/volcano_regions.gpkg", package = "regional"))
  volcano$dis = reg_distinction(vr, volcano, sample_size = 0.5)

  mean(reg_dis$dis)

  plot(volcano)
  plot(vect(vr), add = TRUE)
  plot(volcano)
  plot(volcano["dis"], add = TRUE)
}
}
