% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_distinction.R
\name{reg_distinction}
\alias{reg_distinction}
\title{Title}
\usage{
reg_distinction(region, raster, dist_fun = "euclidean", sample_size = 20)
}
\arguments{
\item{sample_size}{}
}
\value{

}
\description{
Title
}
\examples{
library(terra)
library(sf)
volcano = rast(system.file("raster/volcano.tif", package = "supercells"))
vr = read_sf(system.file("regions/volcano_regions.gpkg", package = "laland"))
reg_dis = reg_distinction(vr, volcano)

mean(reg_dis$dis)

plot(volcano)
plot(vect(vr), add = TRUE)
plot(volcano)
plot(reg_dis["dis"], add = TRUE)
}
