% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_inhomogeneity.R
\name{reg_inhomogeneity}
\alias{reg_inhomogeneity}
\title{Title}
\usage{
reg_inhomogeneity(region, raster, dist_fun = "euclidean", sample_size = 1)
}
\arguments{
\item{sample_size}{}
}
\value{

}
\description{
Title
}
\examples{
\dontrun{
  library(terra)
  library(sf)
  volcano = rast(system.file("raster/volcano.tif", package = "supercells"))
  vr = read_sf(system.file("regions/volcano_regions.gpkg", package = "laland"))
  reg_inh = reg_inhomogeneity(vr, volcano, sample_size = 0.5)

  mean(reg_inh$inh)

  plot(volcano)
  plot(vect(vr), add = TRUE)
  plot(volcano)
  plot(reg_inh["inh"], add = TRUE)
}
}
