% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_isolation.R
\name{reg_isolation}
\alias{reg_isolation}
\title{Title}
\usage{
reg_isolation(region, raster, dist_fun = "euclidean", sample_size = 20)
}
\arguments{
\item{sample_size}{}
}
\value{

}
\description{
Title
}
\examples{
library(terra)
library(sf)
volcano = rast(system.file("raster/volcano.tif", package = "supercells"))
vr = read_sf(system.file("regions/volcano_regions.gpkg", package = "laland"))
reg_inh = reg_isolation(vr, volcano)

mean(reg_inh$iso)

plot(volcano)
plot(vect(vr), add = TRUE)
plot(volcano)
plot(reg_inh["iso"], add = TRUE)
}
