% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional-package.R
\docType{package}
\name{regional-package}
\alias{regional}
\alias{regional-package}
\title{regional: Simulation and evaluation of multi-candidate district elections under various circumstances}
\description{
Simulation and evaluation of multi-candidate district elections under various circumstances. Seat allocation under common allocation methods, vote share thresholds, and accounting for already allocated seats in multi-phase seat allocation (e.g., when allocation national seats to compensate for disproportionality in seats awarded in districts)
}
\author{
\strong{Maintainer}: Karel Kroeze \email{k.a.kroeze@utwente.nl}

}
\keyword{internal}
