% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disproportionality.R
\name{seat_difference}
\alias{seat_difference}
\title{Absolute seat difference}
\usage{
seat_difference(
  votes,
  seats,
  baseline_seats_fn = function(votes) allocate_seats_nl(length(votes), votes)
)
}
\arguments{
\item{votes}{vector of votes}

\item{seats}{vector of allocated seats}

\item{baseline_seats_fn}{function to calculate baseline seat allocation to
compare with. Should take a vector of vote counts as the sole argument. The
default, a thin wrapper around \code{allocate_seats_nl}, follows seat allocation
for the Dutch Parliament.}
}
\value{
integer sum of absolute seat differences
}
\description{
Absolute difference in seats from a known baseline seats allocation method.
}
\details{
Calculates seat allocation under the baseline, then returns the sum of absolute
seat differences. The sum is divided by two, as each seat gained by one party,
is always a seat lost by another party.
}
